shivam_bills/requirements.txt
asgiref==3.2.7
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==2.1.15
django-jet==1.0.8
django-nested-admin==3.3.0
django-simple-history==2.9.0
django-solo==1.1.3
html5lib==1.0.1
Pillow==7.1.1
pycparser==2.20
Pyphen==0.9.5
python-monkey-business==1.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
